requirements.txt
#TODO: update package versions
numpy
pandas
pyEDFlib
pyserial
texttable
sphinx
sphinx-rtd-theme
influxdb-client
reactivex
pytest
